requirements.txt
pandas==2.0.3
pyyaml==6.0
Pillow==10.0.0
scikit-learn==1.3.0
numpy==1.24.3
tensorflow==2.7.0
keras==2.7.0
datetime==5.1
Flask==2.3.2